requirements.txt
numpy==1.24.3
PyYAML==6.0
torch==2.0.1
torchvision==0.15.2
scikit-image==0.20.0
scikit-learn==1.2.2
umap-learn==0.5.3
matplotlib==3.7.1
tqdm==4.65.0
pillow==9.5.0
loguru==0.7.0
pandas==2.0.1